Admin: Move ActionAuthenticationMixin to dedicated authorization.py
