+++ moto/core/__init__.py
@@ -1,6 +1,6 @@
 from .models import DEFAULT_ACCOUNT_ID  # noqa
 from .models import patch_client, patch_resource  # noqa
-from .responses import ActionAuthenticatorMixin
+from .authorization import ActionAuthenticatorMixin
 
 set_initial_no_auth_action_count = (
     ActionAuthenticatorMixin.set_initial_no_auth_action_count
